splits the dense tensor, test=develop
