+++ paddle/fluid/framework/variable.h
@@ -72,7 +72,7 @@ class Variable {
  private:
   // This method hides type T, so it doesn't appear as a template parameter of
   // Variable.
-  pten::TensorInplaceVersion* InplaceVersionCounter();
+  pten::DenseTensor::InplaceVersion* InplaceVersionCounter();
 
  public:
   void SetInplaceVersionToZero();
@@ -114,8 +114,8 @@ class Variable {
   std::shared_ptr<Placeholder> holder_;
 };
 
-inline pten::TensorInplaceVersion* Variable::InplaceVersionCounter() {
-  pten::TensorInplaceVersion* version_counter_ptr(nullptr);
+inline pten::DenseTensor::InplaceVersion* Variable::InplaceVersionCounter() {
+  pten::DenseTensor::InplaceVersion* version_counter_ptr(nullptr);
   if (IsType<framework::LoDTensor>()) {
     version_counter_ptr =
         &GetMutable<framework::LoDTensor>()->InplaceVersionCounter();
